# Revise `Delegate` to use `std::function`