--- a/Sming/Arch/Esp8266/app.mk
+++ b/Sming/Arch/Esp8266/app.mk
@@ -8,6 +8,7 @@
 LIBS += \
 	microc \
 	microgcc \
+	stdc++ \
 	hal \
 	phy \
 	pp \