--- a/Sming/Core/Debug.cpp
+++ b/Sming/Core/Debug.cpp
@@ -20,7 +20,7 @@ void DebugClass::initCommand()
 {
 #if ENABLE_CMD_EXECUTOR
 	commandHandler.registerCommand(CommandDelegate(F("debug"), F("New debug in development"), F("Debug"),
-												   commandFunctionDelegate(&DebugClass::processDebugCommands, this)));
+												   CommandFunctionDelegate(&DebugClass::processDebugCommands, this)));
 #endif
 }