--- a/Sming/Core/Network/Mqtt/MqttPayloadParser.h
+++ b/Sming/Core/Network/Mqtt/MqttPayloadParser.h
@@ -12,7 +12,7 @@
 
 #pragma once
 
-#include "Delegate.h"
+#include <esp_systemapi.h>
 #include "mqtt-codec/src/message.h"
 
 /** @defgroup   mqttpayload Provides MQTT payload parser