--- a/Sming/Core/Network/TcpConnection.h
+++ b/Sming/Core/Network/TcpConnection.h
@@ -23,7 +23,6 @@
 
 #include "WiringFrameworkDependencies.h"
 #include "IPAddress.h"
-#include "Delegate.h"
 
 #define NETWORK_DEBUG