--- a/Sming/Core/SmingCore.h
+++ b/Sming/Core/SmingCore.h
@@ -12,8 +12,6 @@
 
 #define SMING_VERSION "3.8.0" // Major Minor Sub
 
-#include <functional>
-
 #include "gdb/gdb_hooks.h"
 #include "WiringFrameworkIncludes.h"