--- a/Sming/Platform/WifiEvents.h
+++ b/Sming/Platform/WifiEvents.h
@@ -13,7 +13,6 @@
 
 #pragma once
 
-#include "Delegate.h"
 #include "WString.h"
 #include "IPAddress.h"