--- a/Sming/Services/CommandProcessing/CommandDelegate.h
+++ b/Sming/Services/CommandProcessing/CommandDelegate.h
@@ -19,10 +19,13 @@
 /** @brief  Command delegate function
  *  @param  commandLine Command line entered by user at CLI, including command and parameters
  *  @param  commandOutput Pointer to the CLI print stream
- *  @note   commandFunctionDelegate defines the structure of a function that handles individual commands
+ *  @note   CommandFunctionDelegate defines the structure of a function that handles individual commands
  *  @note   Can use standard print functions on commandOutput
  */
-typedef Delegate<void(String commandLine, CommandOutput* commandOutput)> commandFunctionDelegate;
+typedef Delegate<void(String commandLine, CommandOutput* commandOutput)> CommandFunctionDelegate;
+
+/** @deprecated Use `CommandFunctionDelegate` instead */
+typedef CommandFunctionDelegate commandFunctionDelegate SMING_DEPRECATED;
 
 /** @brief  Command delegate class */
 class CommandDelegate
@@ -36,13 +39,13 @@ public:
      *  @param  reqGroup The command group to which this command belongs
      *  @param  reqFunction Delegate that should be invoked (triggered) when the command is entered by a user
      */
-	CommandDelegate(String reqName, String reqHelp, String reqGroup, commandFunctionDelegate reqFunction);
+	CommandDelegate(String reqName, String reqHelp, String reqGroup, CommandFunctionDelegate reqFunction);
 	~CommandDelegate();
 
 	String commandName; ///< Command name
 	String commandHelp; ///< Command help
 	String commandGroup; ///< Command group
-	commandFunctionDelegate commandFunction; ///< Command Delegate (function that is called when command is invoked)
+	CommandFunctionDelegate commandFunction; ///< Command Delegate (function that is called when command is invoked)
 
 private :
 	CommandDelegate();