--- a/samples/Basic_Delegates/app/callbacks.cpp
+++ b/samples/Basic_Delegates/app/callbacks.cpp
@@ -1,4 +1,3 @@
-
 #include "callbacks.h"
 
 int testCounter;