… Sming/System/include/m_printf.h
… samples/Arducam/app/ArduCamCommand.cpp
… samples/Basic_Delegates/app/application.cpp
… samples/Basic_Delegates/app/callbacks.h
… samples/Basic_Delegates/app/speed.cpp
… samples/Basic_Serial/app/SerialReadingDelegateDemo.cpp
… samples/Basic_Serial/app/application.cpp
… samples/Basic_Serial/include/SerialReadingDelegateDemo.h
… samples/Basic_Servo/app/application.cpp
… samples/Basic_rBoot/app/application.cpp
… samples/CommandProcessing_Debug/app/ExampleCommand.cpp
… samples/Websocket_Client/app/application.cpp