Trying to use both `std::function` and `Delegate` causes issues, and they seem to do the same job anyway.

Further to #1337, #1732 and #1733, this PR is a rewrite of the `Delegate` class based on `std::function`.

__Improvements__

* Full compatibility with `std::function`, `std::bind` and lambdas
* `std::bind` is not required for standard callbacks, so existing code does not require modifying
* When invoking class methods, Deprecate uses lambda functions instead of `std::bind`
* No heap usage

__Call types__

* `Basic_Delegates` updated with speed testing to show performance of various call methods with a single parameter. Indicative results on an Esp8266:

	Method | CPU cycles per call | current Delegate
	--- | --- | ---
	Callback | 23 | 23
	Delegate (function) | 45 | 46
	Delegate (lambda) | 43 | -
	Delegate (method) | 51 | 51
	Delegate (bind) | 51 | -

* Note that `std::bind` _may_ also use the heap, and there is anecdotal evidence that type inferences can sometimes go wrong. It's generally safer to use lambdas.
* `std::bind` is only required to bind callbacks to functions or class methods where parameter lists differ.

__Notes__

* Reference counting is not supported any more
* Clearing a Delegate must be done using `nullptr`, not `0` or `NULL`
* `stdc++` library has been added for Esp8266 builds as it may be required for placeholders when building with optimisations disabled

__Deprecations__

* `TimerDelegateStdFunction` not required, just use `TimerDelegate` instead
* Rename `commandFunctionDelegate` to `CommandFunctionDelegate`
* Use `HardwareSerial::onDataReceived` instead of `setCallback`

__Additional changes__

* Replace `std::function` types with `Delegate`
* Replace `std::bind` with Delegate, simpler and uses lambdas
* Remove redundant declarations and includes