@verybadsoldier 2019-06-29
I cannot test it against my application code as I would have to port to new API first but I am sure I am fine with just `std::function`. And I basically think it is a good step forward.