@mikee47 2019-06-29
It would be helpful to see the warnings, perhaps work through them here and we can use it to improve the migration guide.