@verybadsoldier 2019-06-29
Output is like this:
```
+ Libraries/RGBWWLed/RGBWWLedColor.cpp
In file included from Libraries/RGBWWLed/RGBWWTypes.h:3:0,
                 from Libraries/RGBWWLed/RGBWWLed.h:40,
                 from Libraries/RGBWWLed/RGBWWLedColor.cpp:8:
Libraries/RGBWWLed/../../SmingCore/SmingCore.h:2:0: error: ignoring #pragma warning  [-Werror=unknown-pragmas]
 #pragma warning "Please update to #include <SmingCore.h> - see https://github.com/SmingHub/Sming/wiki/Migrating-from-Sming-v3.8.x-to-v4"
 ^
In file included from Libraries/RGBWWLed/RGBWWconst.h:11:0,
                 from Libraries/RGBWWLed/RGBWWTypes.h:4,
                 from Libraries/RGBWWLed/RGBWWLed.h:40,
                 from Libraries/RGBWWLed/RGBWWLedColor.cpp:8:
Libraries/RGBWWLed/../../SmingCore/SmingCore.h:2:0: error: ignoring #pragma warning  [-Werror=unknown-pragmas]
 #pragma warning "Please update to #include <SmingCore.h> - see https://github.com/SmingHub/Sming/wiki/Migrating-from-Sming-v3.8.x-to-v4"
 ^
In file included from Libraries/RGBWWLed/RGBWWconst.h:16:0,
                 from Libraries/RGBWWLed/RGBWWTypes.h:4,
                 from Libraries/RGBWWLed/RGBWWLed.h:40,
                 from Libraries/RGBWWLed/RGBWWLedColor.cpp:8:
Libraries/RGBWWLed/../../SmingCore/SmingCore.h:2:0: error: ignoring #pragma warning  [-Werror=unknown-pragmas]
 #pragma warning "Please update to #include <SmingCore.h> - see https://github.com/SmingHub/Sming/wiki/Migrating-from-Sming-v3.8.x-to-v4"
 ^
In file included from Libraries/RGBWWLed/RGBWWLedColor.h:10:0,
                 from Libraries/RGBWWLed/RGBWWLed.h:43,
                 from Libraries/RGBWWLed/RGBWWLedColor.cpp:8:
Libraries/RGBWWLed/../../SmingCore/SmingCore.h:2:0: error: ignoring #pragma warning  [-Werror=unknown-pragmas]
 #pragma warning "Please update to #include <SmingCore.h> - see https://github.com/SmingHub/Sming/wiki/Migrating-from-Sming-v3.8.x-to-v4"
 ^
```
Looks clear to me but I want to first get rid of two hacks I did for my 3.8.0 version (includes this std::function topic) before going forward further.