@mikee47 2019-07-02
Re. issue https://github.com/verybadsoldier/esp_rgbww_firmware/pull/1#issuecomment-507046790, the 'undefined placeholder` problem. This is the sort of error the linker throws up:

```
r:/sandboxes/esp_rgbww_firmware/out/Esp8266/debug/lib/clib-App-d368dbc221bcd658e6ecbdb532e2b09e.a(webserver.o): In function `ApplicationWebserver::onIndex(HttpRequest&, HttpResponse&)':
r:\sming-v4\sming\core\data/arduinojson.h:86: undefined reference to `std::placeholders::_2'
r:\sming-v4\sming\core\data/arduinojson.h:86: undefined reference to `std::placeholders::_1'
```

This is only evident on the Windows Esp8266 compiler, linux is OK.

The problem manifests when compiling with `-Og` or `-O0` so suspect there's a piece of RTL code missing from somewhere. And yes, if I add `stdc++` to LIBS it links fine.