@mikee47 2019-07-02
Second point. The existing `Delegate` is reference-counted. Or, more precisely, the 'caller' object which the `Delegate` wrapper class contains gets copied onto the heap and ends up being destroyed automatically when all the Delegates referencing it are destroyed.

The advantage of reference counting is we can treat the Delegate almost like a function pointer without doing anything special, like this (from `System.cpp`):

```
void SystemClass::onReady(SystemReadyDelegate readyHandler)
{
	if(readyHandler) {
		auto handler = new SystemReadyDelegate(readyHandler);
		queueCallback(
			[](uint32_t param) {
				SystemClass::state = eSS_Ready;
				auto handler = reinterpret_cast<SystemReadyDelegate*>(param);
				(*handler)();
				delete handler;
			},
			reinterpret_cast<uint32_t>(handler));
	}
}
```

It does make the overhead of such calls clearer. Note that when I wrote this I should have used
 `readyHandler->increase()` and `readyHandler->decrease()`, but because I didn't the code works exactly the same using `std::function`.

It also seems that `bind` can do heap allocation - https://stackoverflow.com/questions/46163607/avoid-memory-allocation-with-stdfunction-and-member-function. Because this all happens at run-time, and is probably compiler dependent, it'll need some investigation to find out what the downsides are going to be.