@mikee47 2019-07-02
@slaff I kind of threw this PR together as a 'what if', so is a bit more drastic than the approach in https://github.com/SmingHub/Sming/issues/1337#issuecomment-374951893. I'm OK with it though!

Porting `esp_rgbww_firmware` was a useful exercise to establish how much work migration would require. As far as delegates were concerned it's fairly straightforward. (ArduinoJson definitely took more work, maybe we could add V5 back in as an option to ease things a bit?)

This PR eliminates Delegate completely, and removes `TimerDelegateStdFunction` instead of deprecating it. I've not renamed anything, so still using the term `Delegate`, but it actually works better for me than, say, `StdFunction` as it ably describes _what_ it does, rather than _how_.

One thing I considered was retaining `Delegate` as an alias for `std::function`, perhaps retaining `Delegate.h` for that which would also include <functional> and <placeholders>. If we're going to keep the term 'delegate' then that might work better. I could also revert a bunch of changes which replaced `Delegate` with `std:;function`. What do you think?

I'll need to edit the top comment to summarise impact before merging.