@mikee47 2019-07-02
I'm also wondering if there's a way we can avoid the need to explicitly use `std::bind`, so we could keep doing this: 

`ntpcp = new NtpClient("pool.ntp.org", 30, NtpTimeResultDelegate(&ntpClientDemo::ntpResult, this));`

instead of having to do this:
`ntpcp = new NtpClient("pool.ntp.org", 30, std::bind(&ntpClientDemo::ntpResult, this, _1, _2));`