@slaff 2019-07-03
> One thing I considered was retaining Delegate as an alias for std::function, perhaps retaining Delegate.h for that which would also include and . If we're going to keep the term 'delegate' then that might work better. I could also revert a bunch of changes which replaced Delegate with std:;function. What do you think?

Yes, why not.