@slaff 2019-07-04
@mikee47  Can you help me understand better the table below?  
- current Delegate ?= the old Delegate mechanism ?
- callback ?= calling anonymous functions/lambdas ?

	Method | time (ns) per call | current Delegate
	--- | --- | ---
	Callback | 275 | -
	Delegate (function) | 663 | 538
	Delegate (method) | 738 | 600
	Delegate (bind) | 738 | -


If I have understood the results it seems that the old Delegate mechanism is faster and the only real disadvantage is that it does not support anonymous functions. Would it be better to add support for lambdas in the old Delegates instead of replacing it completely with the new Delegate suggested in this PR? I am looking for pro and contra arguments :)