@mikee47 2019-07-04
I've rewritten the testing code to show figures in CPU cycles, and add a lambda test case. I also had a suspicion the optimiser was messing things up, that should now be sorted.

Method | CPU cycles per call | current Delegate
--- | --- | ---
Callback | 23 | 23
Delegate (function) | 45 | 46
Delegate (lambda) | 43 | -
Delegate (method) | 51 | 51
Delegate (bind) | 51 | -

To test on a different build just need to comment-out the lambda and std::bind tests.

Hopefully this is now a no-brainer!

I'll also just point out that the old Delegate puts the callback function wrapper on the heap, so for every registered callback there's a RAM cost (and another possible source of heap fragmentation). Using lambdas means that the code is created in the calling context (usually flash, but if the caller is in IRAM then the lambda will be too).

In addition to captured lambdas, the old Delegate doesn't support std::bind either. Because the new Delegate is based on std::function we can guarantee compatibility with all the standard uses.

I think the key point here is that if speed is important, use a function callback, but for flexibility use a Delegate.